requirements.txt
Django
django-ninja
django-sesame
